// src/interfaces/pokemon/pokemon.interface.ts
// Generated by https://quicktype.io

import { GameIndex } from "./game.interface";
import { Move } from "./move.interface";
import { Sprites } from "./sprite.interface";

export interface Pokemon {
  abilities: Ability[];
  base_experience: number;
  forms: Species[];
  game_indices: GameIndex[];
  height: number;
  held_items: any[];
  id: number;
  is_default: boolean;
  location_area_encounters: string;
  moves: Move[];
  name: string;
  order: number;
  past_abilities: any[];
  past_types: any[];
  species: Species;
  sprites: Sprites;
  stats: Stat[];
  types: Type[];
  weight: number;
}

export interface PokemonShort {
  id: string | number;
  name: string,
  types: string,
  abilities: string,
  height: number,
  sprites: string,
  region: string,
}


export interface Ability {
  ability: Species;
  is_hidden: boolean;
  slot: number;
}

export interface Species {
  name: string;
  url: string;
}

export interface Stat {
  base_stat: number;
  effort: number;
  stat: Species;
}

export interface Type {
  slot: number;
  type: Species;
}